refactor: combine ui component exports to single index file
