+++ src/app/(app)/account/UpdateEmailCard.tsx
@@ -7,8 +7,7 @@ import { toast } from 'sonner';
 
 import { updateUser } from '@/lib/actions/users';
 
-import { Button } from '@/components/ui/button';
-import { Input } from '@/components/ui/input';
+import { Button, Input } from '@/components/ui';
 
 export default function UpdateEmailCard({ email }: { email: string }) {
   const [state, formAction] = useFormState(updateUser, {
