+++ src/app/(app)/settings/page.tsx
@@ -3,7 +3,7 @@
 import React from 'react';
 import { useTheme } from 'next-themes';
 
-import { Button } from '@/components/ui/button';
+import { Button } from '@/components/ui';
 
 export default function Page() {
   const { setTheme } = useTheme();
