+++ src/components/auth/SignOutBtn.tsx
@@ -1,10 +1,11 @@
 'use client';
 
 import React from 'react';
-import { Button } from '../ui/button';
 import { useFormStatus } from 'react-dom';
 
 import { signOutAction } from '@/lib/actions/users';
+
+import { Button } from '@/components/ui';
 
 export default function SignOutBtn() {
   return (
